Restructured code
This commit restructures the code. Almost all code was stored in ws.py
before so it became kind of a mess. Now the code is split up in multiple
files.
