+++ bayesianfuncs.py
@@ -1,0 +1,25 @@
+from bayesian.classify import Classify
+from bayesian.learn import Learn
+from parsing import fetch_title_from_msg_content
+
+
+def bayesian_score(title):
+    try:
+        c = Classify()
+        c.validate(["", "", title, "good", "bad"])
+        output = c.execute()
+        return output
+    except:
+        return 0.1
+
+
+def bayesian_learn_title(message_content, doctype):
+    try:
+        bayesian_learn = Learn()
+        bayesian_learn.file_contents = fetch_title_from_msg_content(message_content)
+        bayesian_learn.count = 1
+        bayesian_learn.doc_type = doctype
+        bayesian_learn.execute()
+        return True
+    except:
+        return False
