+++ utcdate.py
@@ -1,0 +1,5 @@
+from datetime import datetime
+
+
+class UtcDate:
+    startup_utc_date = datetime.utcnow()
